requirements.txt
polib==1.2.0
openai==1.55.3
python-dotenv==1.0.1